refactor(server): factor out metrics code into metrics.{h,cc}
Move the Metrics aggregation struct and Prometheus rendering out of
server_family.{h,cc} into a dedicated metrics.{h,cc} translation unit.

- Move Metrics, ReplicationMemoryStats, LoadingStats, PeakStats to metrics.h
- Convert PrintPrometheusMetrics() into Metrics::Print(); move Merge() and
  InitFromThread() to metrics.cc
- Promote GetDelayMs/ReadProcStats to header-declared free functions so both
  server_family.cc (INFO command) and metrics.cc can share them
- Add metrics.cc to dragonfly_lib

Pure code movement; no behavior change.
